requirements.txt
branca==0.7.2
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
contourpy==1.3.0
cycler==0.12.1
db-dtypes==1.3.0
folium==0.17.0
fonttools==4.53.1
geopandas==1.0.1
google-api-core==2.20.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-cloud-bigquery==3.25.0
google-cloud-core==2.4.1
google-crc32c==1.6.0
google-resumable-media==2.7.2
googleapis-common-protos==1.65.0
grpcio==1.66.1
grpcio-status==1.66.1
idna==3.10
Jinja2==3.1.4
kiwisolver==1.4.7
MarkupSafe==2.1.5
matplotlib==3.9.2
numpy==2.1.1
oauthlib==3.2.2
packaging==24.1
pandas==2.2.3
pandas-gbq==0.23.1
pillow==10.4.0
proto-plus==1.24.0
protobuf==5.28.2
pyarrow==17.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydata-google-auth==1.8.2
pyogrio==0.9.0
pyparsing==3.1.4
pyproj==3.6.1
python-dateutil==2.9.0.post0
pytz==2024.2
requests==2.32.3
requests-oauthlib==2.0.0
rsa==4.9
seaborn==0.13.2
shapely==2.0.6
six==1.16.0
tzdata==2024.1
urllib3==2.2.3
xyzservices==2024.9.0
